Added printDealer1stHand function and fixed lack of garbage collection
is_hitting flag is retained inside of play loop, if player busts or blackjacks the flag remains as garbage collection doesn't have a chance to reset the flag. Set the flag back to false in the beginning if player chooses not to hit to prevent adding a card when choosing not to.